gen4: add append for u-boot_2022.01.bb
The meta-rcar-bsp for the v4h is using u-boot 2022.01 and the patch
0001-net-rswtich-disable-GWCA-when-removing-driver.patch became fuzzed.

At the same time meta-rcar-gateway (s4) is using u-boot 2020.10 with
its own revision of the same patch.

In order to be able to apply proper version of the patch, this commit
adds new bbappend with its own directory for patch.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>